requirements/test.txt
#factory_boy==2.7.0
#model_mommy==1.2.6

pytest==3.0.1
pytest-django==3.0.0
pytest-flake8==0.6
pytest-splinter==1.7.6
pytest-sugar==0.7.1
pytest-xdist==1.15.0
